+++ src/index.ts
@@ -1,4 +1,21 @@
-import getSchemaForEndpoint from './functions/get-schema-for-endpoint';
+import {
+  getSchemaForEndpoint,
+  getSchema,
+  validateSchema,
+} from './functions/schema';
+import {
+  getOnchainMetadata,
+  getOnchainMetadataVersion,
+  getCIPstandard,
+} from './functions/metadata';
 
-export { getSchemaForEndpoint };
+export {
+  getSchemaForEndpoint,
+  getSchema,
+  validateSchema,
+  getOnchainMetadata,
+  getCIPstandard,
+  getOnchainMetadataVersion,
+};
+
 export type { paths, components } from './generated-types';